refactor: split lib.rs FFI layer into ffi/ module and resolve neural context TODO
Split the monolithic lib.rs (2396 lines) into engine/src/ffi/ module structure:
- mod.rs: shared helpers, macros (ffi_guard!, ffi_open!, ffi_close!), OwnedVec
- dict.rs, connection.rs, convert.rs, history.rs, romaji.rs: domain FFI
- candidates.rs: unified candidate generation FFI
- session.rs: InputSession FFI with with_history closure pattern
- neural.rs: neural scorer FFI (cfg-gated)
- tests.rs: all FFI tests

Also resolves the neural context TODO:
- Add committed_context field to InputSession, accumulated on each commit
- Add lex_session_committed_context / lex_committed_context_free FFI functions
- Update LeximeInputController.swift to pass actual context for neural candidates
